requirements.txt
google-genai>=1.0.0
pymupdf>=1.24.0
requests>=2.31.0
